+++ Config.h
@@ -4,11 +4,11 @@
 struct Config {
     int numCPUs;
     std::string schedulerType;
-    int quantumCycles;
-    int batchProcessFreq;
-    int minInstructions;
-    int maxInstructions;
-    int delaysPerExec;
+    unsigned long long quantumCycles;
+    unsigned long long batchProcessFreq;
+    unsigned long long minInstructions;
+    unsigned long long maxInstructions;
+    unsigned long long delaysPerExec;
 };
 
 Config loadConfig(const std::string& filePath = "config.txt");
